Доработки по заданию 'Отрисуй меня полностью (часть 1)'
